stop the client from being deleted if there's related data

+++ HNstock/app/Http/Controllers/ClientController.php
@@ -117,6 +117,10 @@ class ClientController extends Controller
      */
     public function destroy(Client $client)
     {
+        if (!$this->repository->isDeletable($client)) {
+            return to_route(route: 'clients.index')->with('error', 'cannot delete client');
+        }
+
         $client->delete();
         return to_route(route: 'clients.index')->with('success', 'client deleted successfully');
     }
